+++ esdt-nft-marketplace/Cargo.toml
@@ -9,7 +9,7 @@ publish = false
 path = "src/marketplace_main.rs"
 
 [dependencies.elrond-wasm]
-version = "0.32.0"
+version = "0.33.0"
 
 [dev-dependencies.elrond-wasm-debug]
-version = "0.32.0"
+version = "0.33.0"
