+++ esdt-nft-marketplace/meta/Cargo.toml
@@ -9,7 +9,7 @@ publish = false
 path = ".."
 
 [dependencies.elrond-wasm]
-version = "0.32.0"
+version = "0.33.0"
 
 [dependencies.elrond-wasm-debug]
-version = "0.32.0"
+version = "0.33.0"
