+++ esdt-nft-marketplace/src/common_util_functions.rs
@@ -1,0 +1,12 @@
+elrond_wasm::imports!();
+
+#[elrond_wasm::module]
+pub trait CommonUtilFunctions {
+    fn get_nft_info(&self, nft_type: &TokenIdentifier, nft_nonce: u64) -> EsdtTokenData<Self::Api> {
+        self.blockchain().get_esdt_token_data(
+            &self.blockchain().get_sc_address(),
+            nft_type,
+            nft_nonce,
+        )
+    }
+}
