+++ seller-contract-mock/meta/Cargo.toml
@@ -11,7 +11,7 @@ authors = ["Dorin Iancu <dorin.iancu@elrond.com>"]
 path = ".."
 
 [dependencies.elrond-wasm]
-version = "0.32.0"
+version = "0.33.0"
 
 [dependencies.elrond-wasm-debug]
-version = "0.32.0"
+version = "0.33.0"
